Extract RestaurantCard component to a separate file
